# docs/specs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "REST API for Stratum — organizations, accounts, billing, and notifications.",
        "title": "Stratum API",
        "contact": {},
        "version": "1.0"
    },
    "basePath": "/api/v1",
    "paths": {
...
        "/me/notifications": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the caller's notifications. Page-based (limit/page) unless a cursor is supplied, in which case it switches to cursor pagination.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "List notifications",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter by organization",
                        "name": "organization_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "filter by channel",
                        "name": "channel",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "max results (default 20, max 100)",
                        "name": "limit",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "description": "page number, ignored once cursor is set (default 1)",
                        "name": "page",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "pagination cursor from a previous response; switches to cursor mode",
                        "name": "cursor",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/messageRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/me/notifications/preferences": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the caller's per-channel/event-type notification preferences.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "List notification preferences",
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/preferenceRecord"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            },
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Creates or updates the caller's preference for a given channel + event type.",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Set a notification preference",
                "parameters": [
                    {
                        "description": "Preference to set",
                        "name": "body",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/upsertPrefRequest"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/preferenceRecord"
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "422": {
                        "description": "validation failed",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/me/notifications/read-all": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Mark all notifications as read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "scope to an organization",
                        "name": "organization_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/me/notifications/stream": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Server-sent events stream; emits a \"notification\" event whenever a new message arrives for the caller. Only mounted when Redis is configured. Capped at 3 concurrent streams per caller.",
                "produces": [
                    "text/event-stream"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Stream notifications (SSE)",
                "responses": {
                    "200": {
                        "description": "text/event-stream connection"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "429": {
                        "description": "too many concurrent streams already open",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/me/notifications/unread-count": {
            "get": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "description": "Returns the number of unread notifications for the caller, optionally scoped to an organization.",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Get unread notification count",
                "parameters": [
                    {
                        "type": "string",
                        "description": "filter by organization",
                        "name": "organization_id",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/Payload"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "unread": {
                                                    "type": "integer"
                                                }
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
        "/me/notifications/{id}/read": {
            "patch": {
                "security": [
                    {
                        "BearerAuth": []
                    }
                ],
                "tags": [
                    "notification"
                ],
                "summary": "Mark a notification as read",
                "parameters": [
                    {
                        "type": "string",
                        "description": "Notification ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "204": {
                        "description": "no content"
                    },
                    "401": {
                        "description": "missing/invalid auth token",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    },
                    "404": {
                        "description": "notification not found",
                        "schema": {
                            "$ref": "#/definitions/Payload"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "Payload": {
            "type": "object",
            "properties": {
                "data": {},
                "next_cursor": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/pagination"
                },
                "status": {
                    "$ref": "#/definitions/Status"
                }
            }
        },
...
        "Status": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "string"
                },
                "details": {},
                "error": {
                    "type": "boolean"
                },
                "message": {
                    "type": "string"
                },
                "request_id": {
                    "type": "string"
                }
            }
        },
...
        "messageRecord": {
            "type": "object",
            "properties": {
                "auth_sub": {
                    "type": "string"
                },
                "body": {
                    "type": "string"
                },
                "channel": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "kind": {
                    "type": "string"
                },
                "organization_id": {
                    "type": "string"
                },
                "payload": {
                    "type": "array",
                    "items": {
                        "type": "integer"
                    }
                },
                "read_at": {
                    "type": "string"
                },
                "sent_at": {
                    "type": "string"
                },
                "status": {
                    "type": "string"
                },
                "subject": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "pagination": {
            "type": "object",
            "properties": {
                "current_page": {
                    "type": "integer"
                },
                "limit": {
                    "type": "integer"
                },
                "offset": {
                    "type": "integer"
                },
                "total_items": {
                    "type": "integer"
                },
                "total_pages": {
                    "type": "integer"
                }
            }
        },
...
        "preferenceRecord": {
            "type": "object",
            "properties": {
                "channel": {
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "event_type": {
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "upsertPrefRequest": {
            "type": "object",
            "required": [
                "channel",
                "event_type"
            ],
            "properties": {
                "channel": {
                    "type": "string"
                },
                "enabled": {
                    "type": "boolean"
                },
                "event_type": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "BearerAuth": {
            "description": "Bearer token issued by Supabase auth.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}